grok-telegram-bot/requirements.txt
python-telegram-bot==20.7
openai==1.6.1
python-dotenv==1.0.0
httpx==0.25.2
